requirements.txt
setuptools
ipykernel==6.23.1
tokenizers==0.19.1
torch==2.3.0
tqdm==4.65.0
transformers==4.44.0
datasets==2.12.0
beautifulsoup4==4.12.2
hydra-core==1.3.2
tensor-parallel==1.2.4
docker==7.1.0
dspy-ai==2.4.12
omegaconf==2.3.0
optuna==3.6.1
wandb
